Add descriptor files to resources
